Refactor Evaluator tests into test class
