+++ google-cloud-debugger/test/google/cloud/debugger/breakpoint/evaluator/immutable_classes_test.rb
@@ -14,19 +14,8 @@
 
 
 require "helper"
-require_relative "expression_test_helper"
 
-describe Google::Cloud::Debugger::Breakpoint::Evaluator do
-  let(:evaluator) { Google::Cloud::Debugger::Breakpoint::Evaluator }
-
-  before do
-    if ENV["GCLOUD_TEST_COVERAGE_DEBUGGER_TIMEOUT"]
-      # Have to set it here because configure gets reset by some tests.
-      eval_time_limit = Float ENV["GCLOUD_TEST_COVERAGE_DEBUGGER_TIMEOUT"]
-      Google::Cloud::Debugger.configure.evaluation_time_limit = eval_time_limit
-    end
-  end
-
+describe Google::Cloud::Debugger::Breakpoint::Evaluator, :evaluator do
   describe "Bignum" do
     it "allows #multiplication" do
       expression_must_equal "123123123123123123123 * 3",
